Write a benchmark

Write a benchmark

benchmark_write(
  benchmark,
  name,
  branch,
  block = NA,
  iteration = NA,
  append = TRUE
)

Arguments

benchmark

The result of bench::mark(), with iterations = 1.

name

The name of the benchmark.

branch

A character vector of length one to indicate the git branch (i.e. commit, tag, branch etc) of the benchmarking.

block

All branches that appear once in a block.

iteration

An integer indicating to which iteration the benchmark refers to. Multiple iterations within a block always benchmark the same branch.

append

Whether to append the result to the file or not.

Value

Character vector of length one with path to the record written (invisibly).