Turn raw benchmark results into text and figures

Turn raw benchmark results into text and figures

benchmarks_analyze(
  refs = c(ref_get_or_fail("GITHUB_BASE_REF"), ref_get_or_fail("GITHUB_HEAD_REF")),
  ci = 0.95
)

Arguments

refs
The names of the branches for which analysis should be created.

ci
The confidence level, defaults to 95%.

Details

Creates two side effects:

Density plots for each element in refs are written to touchstone/plots.

A text explaining the speed diff is written to touchstone/pr-comment/info.txt for every registered benchmarking expression.