The script for benchmarking

The script that contains the code which executes the benchmark. It is typically called with with_touchstone_lib().

Requirements

A touchstone script must:

install all versions of the benchmarked repository with refs_install().

create benchmarks with one or more calls to benchmark_run_ref().

produce the artifacts required in the GitHub workflow with benchmarks_analyze().